Create obtaining state limits
This file was used to obtain the state limits for the residue dihedral states and water orientations. The histogram bin number and histogram smoothing parameters need to be adjusted for certain residues to ensure the gaussian fitting is accurate. Visual inspection of the gaussian fit and the intersect output is necessary to ensure the state limits are obtained correctly.