Create Obtaining Water Distributions
This code obtains the distributions for the waters used in the analysis. Water pockets are defined and checked for water occupation at every frame of the simulation. If water occupies the pocket, the dipole moment is converted into spherical coordinates. If no water occupies the pocket, we assign an artificial empty state.